Improve Deer Valler status reporting
'Closed for the day' lifts now appear as 'Closed' and not 'Scheduled'

Also coerce now reports questionable status

If liftie does not recognize the status it report is as schedules and
logs the warning through debug module.

DEBUG=liftie.coerce

needs to be set to make such errors appear in the log.