Support dataUrl in resort descriptor
In some cases Liftie retrieves lifts status from a different place, or
in a different format than the page we want to send users to.
If dataUrl is specified Liftie will request data from that URL.
Standard url still needs to be provided so that we know where to send
users clicking on official resort links.