Persistent cache for status to limit requests on restart
Status data is now kept in 'node-dirty' database.
Location of database file is determined by looking up RUN_DIR
environment variable (should be set to /var/run/liftie or similar)

node-dirty database is using append only log format, it needs to be
trimmed from time to time. In our case it's safe to remove the file if
it grows to be too big.