Update Vail parser to include following through the waitingroom
Checks if the page contains the document.location.href in a script tag, which is present on the page that is loaded when being sent to the waiting room. This part is mostly so that no changes are required if they choose to remove the waitingroom function. If a waitingroom is detected, we get the content of the page that is referenced in that tag and parse that data, returning the relevant lift statuses.

Two changes to the core:
lib/lifts/request.js: added agent(). This keeps session info to ensure cookies are kept when being kicked around by the waitingroom. Not sure if this will have other implications.
lib/lifts/request.js: added a Promis.resolve. Adding another request within the Vail parser means that a Promise is returned there. Adding the resolve means that the function will wait for the Promise to come through, if it is a promise at all, or continue nevertheless if it isn't a promise. This should ensure no other resorts are broken by this change.

Closes #36